separate version from build
