+++ increaseEventEmitterLimit.mjs
@@ -1,0 +1,5 @@
+// increase the event emitter limit
+
+import { EventEmitter } from 'events';
+
+EventEmitter.defaultMaxListeners = 25;
